Increased technical debt but extracted graph behavior
Need to break down and remove the god-like MaestroBuilder
class - perhaps keeping retaining walker builder.
