+++ src/Extension/Git/Command/GitTagCommand.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Maestro\Extension\Git\Command;
+
+use Symfony\Component\Console\Command\Command;
+
+class GitTagCommand extends Command
+{
+    protected function configure()
+    {
+    }
+
+    public function getAliases()
+    {
+    }
+}
